Move components in the root level to the proper components directory
