+++ src/components/ErrorBoundary.tsx
@@ -1,5 +1,5 @@
 import { Component, ErrorInfo, ReactNode } from 'react'
-import { logUi } from './debug'
+import { logUi } from '../debug'
 
 interface Props {
   children?: ReactNode
